Adds a new lattice_particles model
Starts the addition of the lattice_particles model for making
Monte-Carlo simulations of particles on generic 2 and 3-D lattices.
Proprely integrates the resulting files to CMake, yielding something
that compiles.

So far, the files are the same as the default model library, in order to
have something to get started from.

Minor details: removes trailing spaces from some files, slightly updates
some of the READMEs with more precise instructions.