Adds the geometry functions to lattice_particles
Adds geometry functions, mostly copied from the fields library.

Does not include the creation of the contact map, as we will keep it
flattened and hash the particle properties into a single integer to
determine contact energies.

Adds the functions to hash particle properties into contact index in the
vector_utils library.

Changes the get_neighbours function so that it updates a fixed-sized
array rather than allocating memory to a vector, which should lead to a
speedup.